libtcmu: carve out store calls into separate source
We would be adding asynchronous version of emulated
commands such as compare and write, write verify etc.,
so let's not put all of that into one source and it
up.
